Simplify ALU output multiplexing.
This vastly reduces logic usage, and also helps untangle the critical
path by linking I/O address generation to a shorter portion of it.

A mirroring commit is present on the "main" branch - the full shifter
has about a 50-LUT cost, and reduces average emulation length from 160
clocks to 80 clocks.